[feature #146368967] use pascalcase for file naming
