+++ client/actions/documentActions.js
@@ -2,7 +2,7 @@
  * @desc Handles all actions relating to Documents on the App
  */
 import axios from 'axios';
-import * as types from './actionTypes';
+import * as types from './ActionTypes';
 
 /**
  *
