+++ client/components/common/Header.jsx
@@ -3,9 +3,7 @@ import PropTypes from 'prop-types';
 import { Link, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import AppBar from 'material-ui/AppBar';
-import * as userActions from '../../actions/userActions';
+import * as userActions from '../../actions/UserActions';
 
 class Header extends Component {
   constructor(props) {
@@ -41,7 +39,7 @@ class Header extends Component {
       <div className="navbar-fixed">
         <nav>
           <div className="nav-wrapper">&nbsp;
-            <a href="#!" className="brand-logo"> SHELFDMS </a>
+            <a href="#" className="brand-logo"> SHELFDMS </a>
             <ul className="right hide-on-med-and-down">
               <li><a href="sass.html"><Link style={linkColor} to="/dashboard"> My Dashboard</Link></a></li>
               <li><a href="badges.html"><Link style={linkColor} to="/document">Create a Document</Link></a></li>
