+++ client/components/document/DocumentForm.jsx
@@ -27,19 +27,17 @@ const DocumentForm = ({ document, onSave, onChange, saving }) =>
 
                 <div className="row">
                   <h5 className="col l1 m1 s12">Access:</h5>
-                  <div className="section">
-                    <div className="input-field col l3 m1 s12">
-                      <select
-                        className="browser-default"
-                        value={document.access}
-                        onChange={onChange}
-                      >
-                        <option value="" disabled selected>Choose your option</option>
-                        <option value="private">Private</option>
-                        <option value="public">Public</option>
-                        <option value="role">Role</option>
-                      </select>
-                    </div>
+                  <div className="input-field col l3 m1 s12">
+                    <select
+                      className="browser-default"
+                      value={document.access}
+                      onChange={onChange}
+                    >
+                      <option value="">Choose your option</option>
+                      <option value="private">Private</option>
+                      <option value="public">Public</option>
+                      <option value="role">Role</option>
+                    </select>
                   </div>
                 </div>
               </div>
