+++ client/index.js
@@ -5,11 +5,11 @@ import { render } from 'react-dom';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 import axios from 'axios';
 import jwtDecode from 'jwt-decode';
-import configureStore from './store/configureStore';
-import routes from './routes';
-import initialState from './reducers/initialState';
-import setAuthorizationToken from './utils/authentication';
-import { setCurrentUser } from './actions/userActions';
+import configureStore from './store/ConfigureStore';
+import routes from './Routes';
+import initialState from './reducers/InitialState';
+import setAuthorizationToken from './utils/Authentication';
+import { setCurrentUser } from './actions/UserActions';
 import '../node_modules/materialize-css/dist/js/materialize.min';
 import '../node_modules/materialize-css/dist/css/materialize.min.css';
 import '../node_modules/sweetalert/dist/sweetalert.css';
