+++ client/reducers/documentPaginationReducer.js
@@ -1,5 +1,5 @@
-import * as types from './../actions/actionTypes';
-import initialState from './initialState';
+import * as types from './../actions/ActionTypes';
+import initialState from './InitialState';
 
 export default (state = [initialState.paginatedDocuments], action) => {
   switch (action.type) {
