+++ client/reducers/rootReducer.js
@@ -1,13 +1,13 @@
 import { combineReducers } from 'redux';
-import users from './usersReducer';
-import user from './userReducer';
-import roles from './roleReducer';
-import document from './docReducer';
-import documents from './documentsReducer';
-import message from './messageReducer';
-import isAuth from './authUserReducer';
-import loggedInUserDocuments from './loggedInUserDocumentsReducer';
-import paginatedDocuments from './documentPaginationReducer';
+import users from './UsersReducer';
+import user from './UserReducer';
+import roles from './RoleReducer';
+import document from './DocumentReducer';
+import documents from './DocumentsReducer';
+import message from './MessageReducer';
+import isAuth from './AuthUserReducer';
+import loggedInUserDocuments from './LoggedInUserDocumentsReducer';
+import paginatedDocuments from './DocumentPaginationReducer';
 
 const rootReducer = combineReducers({
   users,
