+++ package.json
@@ -10,7 +10,7 @@
     "lint:watch": "npm run lint -- --watch",
     "start:D": "webpack-dev-server",
     "start:dev": "nodemon ./bin/www",
-    "test:client": "NODE_ENV=test istanbul cover _mocha --require ./client/tests/testSetup.js -R client/tests/**/*.js",
+    "test:client": "NODE_ENV=test istanbul cover _mocha --require ./client/tests/TestSetup.js -R client/tests/**/*.js",
     "test:server": "NODE_ENV=test ./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha --report lcovonly -- -R spec ./server/tests --timeout 20000",
     "script": "sequelize db:migrate:undo:all && sequelize db:migrate && sequelize db:seed:all",
     "seed": "npm run migrate && sequelize db:seed:all",
